+++ backend/src/main/java/ca/mcgill/ecse321/Mar1HotelSystem/dto/CustomerDto.java
@@ -8,7 +8,7 @@ public class CustomerDto extends AccountDto {
     }
 
     // Customer constructor requiring firstName, lastName, email, phoneNumber, password
-    public CustomerDto(String firstName, String lastName, String email, int phoneNumber, String password) {
+    public CustomerDto(String firstName, String lastName, String email, long phoneNumber, String password) {
         super(firstName, lastName, email, phoneNumber, password);
     }
 }
